Removing express in favor of koa
DESCRIPTION:

Moving away from express as it seems like it will be deprecated in the future.
Implemented koa as an alternative. This modifies the way that middleware
is handled so I will need to update the middleware file in the future.
For now, I am removing it as the current middleware is too small to be
separated out.

TEST PLAN:

Run the server to ensure that it works as expected.