Moving away from Graphql
After a lot of consideration, I've decided that I will not continue
using graphql for this project. I think that it will add too much
complexity for the project and ultimately result in a lot of additional
time spent solving graphql related issues without much of a benefit.

Besides that, I think that this application will make much more sense
with a REST API anyways as the fields being queried will always be
needed by the system. I wasn't planning on clients having to specify
every single flag that they want to get information about anyways so the
whole querying aspect of graphql will go to waste.

If I find a use for graphql here, I may go back and add it again.