Moving middleware to its own file and adding CORS middleware
I'm still debating whether or not I should support CORS middleware. I know
that for applications outside of browsers that will access my api, CORS
will not be an issue, however I wonder if I should not support it in case
there is a need for access to flags on the client side. I'll have to think
on it for now, but will be adding it for now.