Fixing lint not working
Turns out that when moving to the monorepo, I needed to move .eslintrc.json
to the root of the monorepo and then add a .eslintrc.json file to the root
of my backend package that extended the functionality at the root. Now I should
be able to get lint rules that are shared across all workspaces working.