Switching to npm from yarn 2 for now
I've looked into npm, pnpm, yarn classic, and yarn modern and have come
to the following conclusions:

1. npm is the most popular package manager and has the most support
2. npm does however suffer from a few issues which includes package duplication
    within nested dependencies
3. pnpm, although fast, installs packages globally which may prove to be
    problematic for developers that are using multiple packages in other
    projects
4. yarn classic is also popular and works just like npm. Yarn classic
    is no longer receiving updates however
5. yarn modern moves node_modules into a local cache. I think this pollutes
    changes but has the added benefit of now having to rely on npm for installs
    and it fixes the duplication issue that npm suffers from in node_modules

For my purposes now, I think that npm is probably the best way to go. I
will probably change my mind and switch back to Yarn 2 in the future especially
as the project grows and I start to add more people to the project. Npm is
simplest and easiest to use for now while still getting updates.