Moving back to Page routing and adding a new page
I found the App routing to still be a bit unstable causing some issues
with loading pages. Occasionally there would be a crash when changing
pages which made me think it was a good idea to move back to the Page
routing system.

I also added some new pages as well as a FlagList component to display
flags using the Mantine UI system. I'm still getting the hang of how
to properly get thigns working, but it's a start.