Added global css file, renaming things
Added a global css file to fix some styling issues. Renamed the sidebar
to navbar and adjusted the mantine components to now start with M. The
reasoning being that I want to be able to keep them separate for now in
case that I want to change things in the future.

I've also gone ahead and changed imports to no longer use the relative
pathing. This is because I want to be able to change the file structure
in the future without having to change all the imports.