Followed the mantine nextjs template and added things
I wasn't aware of how good the mantine nextjs template was and started
to read through it. I ended up adding a few things from it that I liked
and will continue to read through it to understand exactly what's going
on in it.

I added the color scheme toggle, updated the _app.tsx and _document.tsx
files, as well as added a package to handle cookies.