Added an error state for flags as well
Just expirementing with some states for the flags. Will need to figure
out what the best way to organize or visualize this is.

I could for example, keep the states of the current user for each flag,
but that may not be important for some users as they may not pass any of
the flags that are presented to them. Another option is to tie it to the
test user within the accordion, but since the two menus are separate, it
may not be the most intuitive for the user.