Added custom nodes to diagram
I've started to experiment with some custom nodes for the diagram. I've
found that it's not difficult to add custom handles or even supplement
the nodes with MantineUI components. I've gone ahead and added a node that
uses the MantineUI Card component with custom handles.

As for typing, I need to spend a bit more time. I figured out a simple
way to get things working, but I'd like to ensure that types are better
defined.

I also ran into an issue with hydration around the actionIcons used in
the Nav menu for flags. I need to investigate further, but the current
diff fixes the issue.