Fixed typing issues for future node
Spent some time trying to figure out how to best type the nodes. I found
that it was actually a bit more complicated than expected.

There are two parts to the type:
- When then list of nodes is created, the data field for that node needs
to be linked to the type field so that the correct kind of component can
be rendered
- ReactFlow needs to know which component to render based on the type field

At least one of these parts needs to be explicitly defined. The later can
be derived from the prior. I decided to explicitly define the correlation
between the data and type fields. This allows me to automatically create
the ReactFlow required type for custom nodes.

Originally I wrote a hack to allow the type field to be a string. This
after using mantine a bit, I started to get used to how simple it looks.
I opted however to not use it in favor of defining an enum for the type
field. This will make it easier for future refactors when the time comes.
I also utilized typescript's Uncapitalize utility type to make the enum
values look like property types for the ReactFlow nodeTypes object.

There is dummy DropdownNode component that I created to test these new
types, but I haven't made it any different from the CardNode component.

I also updated the tsconfig.json file to better import new custom types.
I might change the name in the future, but I couldn't use @types because
it's a reserved global import from typescript and causes issues.