Separating things but need to learn more
Currently thinking about how I am going to organize things so that I can
easily work with new flags. Since I don't understand how ReactFlow does
its updates, it's making things a bit difficult.

The main idea is that the backend will pass a list of flags to the frontend
yet it probably won't have the positions of the flags in the diagram. I
guess that I could pass that information between both ends, but I don't
know how much more expensive that would be.

I also need to think about how I'm going to color the nodes and edges based
on the current flag. I have yet to implement that and need to think about
how I'm going to do that. It might have an impact on where all the node
handling will exist.