Changing ReactFlow defaults and moving around state
I found that ReactFlow has a few additional fields that I can use to
change the way that it handles inputs, and renders the graph. I've
changed the some snapping and grid settings, but they are currently
hardcoded. I will eventually move the state of these settings into the
parent component so that they can either be stored in a cookie or in
user settings in the backend.

I've also moved the state of the graph into the parent component. This
is because I want to be able to store data about the positions of the
elements in the graph so that users can save it how they like and not
have it reset every time they open the flag. This may also be useful
if I add in collaborative editing, but that's a long way off.

Finally I added some code for start and end nodes. I think that these
nodes will be part of every flag to determine the flow. I may opt-out
of them later and simply use nodes that don't have parents as starting
points and those without children as endpoints.