Created an ActionBar that can add new nodes
Fixed an issue where FlowDiagram was larger than its parent. It should
now sit nicely and will the parent component.

I added an ActionBar which include the ability to add a new node, use
the cursor to move, and a pan button. The new node button adds a new
node the flow diagram while the other buttons are not yet implemented.

Adding of a new node currently requires the position to be known the
moment that the new node button is pressed. I want to instead create a
card that follows the cursor and then when the user clicks, the card is
replaced by the node. This will allow the user to move the cursor around
to find the right position before clicking.

I want to make sure that the card sits "in" the flow diagram. That would
mean covering the card with the ActionBar and other boxes while sitting
on top of the FlowDiagram. I want to avoid setting z-indexes if possible
so I will need to fiddle around with things.