Updated the ActionBar
Instead of having the ActionBar sit at the top of the page, it now sits
as a panel of the FlowDiagram. I've also gone and replaced the controls
for the diagram as they will now be part of the ActionBar.

I've also added a new button from the controls to the ActionBar called
"Fit View" which works the same as the fit view from the controls.