Styling ReactFlow gets messy
Styling for ReactFlow is a bit odd as it uses classnames which cannot be
manipulated by React state. As such, I have to use a hacky solution to
get the styling to work.

This involves creating a stylesheet for ReactFlow and imptorting all the
styles from the styles.css file. I then create a var
in the css stylesheet that I can search for within JS and modify. This
causes some type issues in TypeScript which have to be ignored.

Ultimately, this is a hacky solution that bothers me but seems to be the
only way to accomplish what I want without having to fork ReactFlow and
modify it directly.