Adds selecting of action and a new connector action
When an action from the AcitonBar is clicked, it will become selected.
This is the case for each of the icons except for fit as it does not
change the way the user interacts with the diagram.

The newNode and newEdge actions will have some sort of visual for the
user. When the newNode is selected a temporary node will be visible and
when the newEdge is selected, a visual indicating how the user can
connect nodes will be visible. These have not been implemented yet.

I may move the fit action somewhere else similar to Figma