Moved Nodes into a Node folder, Added CardNode styles
Decided to move everything into a Node folder so that it's easier to
manage.

I've also updated the CardNode to have special styles when it's selected
or not. There are also some styles for passing and failing, yet they are
not connected to anything. I'll need to figure out how to pass that info
down to the CardNode. It might simply be included as part of the data.