Testing an approach to adding colors
I'm testing out a way to set the colors of the flow diagram. I think that
I can pass in a set of records with ids and statuses that determine how
they are meant to be colored, and the the flow diagram can later use
those values to color the nodes.

In this simple version however, I'm manually setting the colors of the
first four nodes that are added to the flow diagram. This is utilizing
the onNewNode callback, but can probably exist in the onNodesChange
function instead to propogate the colors to all nodes that need them.
If the flags are passed into the FlowDiagram, then I may need a useEffect
to color all of the nodes when only the results change.