Modified css import order to override ReactFlow styles
Wow this was a pain. I looked through a ton of different places to find
out how to accomplish this and was stumped for days.

What is the problem?
ReactFlow utilizes pure css classes for styling. This is normally not an
issue as you can override those styles by using new class names either
within CSS or CSS in JS. Mantine uses emotion, a CSS in JS library,
which under the hood prepends all the matine styles and classes to the
beginning of the head tag. This means that any classes that I created
with Mantine would all exist *before* the ReactFlow classes resulting in
none of the new styles overriding the previous ReactFlow ones.

Because of this, I had to figure out how to get the Mantine styles to
load after ReactFlow.

I tried to modify the document and add the html
with the ReactFlow css stylesheet at the beginning, but it would get
overritten by emotion.

I tried adding the ReactFlow styles in JS, but the names of the classes
were all invalid properly names due to them being in kebab case.

Ultimately, the solution was to create a custom emotion cache and force
the mantine styles to not be prepended in the head tag. I will probably
need to look further into cache ordering to ensure that future libraries
do not cause issues as this seems to be a prevelant problem between dev
and prod builds.

I'm honestly suprised it took so long to figure this out. If I took some
more time to read through the Mantine docs, I might have been able to
solve this problem sooner.