Modified behavior of CardNode handles
The CardNode has two handles, top and bottom, that are used to connect
the nodes to other nodes. I want to have a similar effect to Figma where
the handles are only visible when the node is selected or when hovered
over and the connection mode is on.

To do this, I created the first step which allows for the handles to be
visible depending on a prop called showHandles. Here is the way that it
works:

If the handle is disabled, it should never be visible. If the handle is
not disabled, then it should only be visible if the showHandles prop is
true. I will be updating the visibility of handles on select in a future
commit.