Updated cards to show handles on selection
I modified the handles a little bit so that they stick out from the card
a little bit. I'm a fan of the style,  just not of the color so I will
come back to that later.

I also need to modify the way that handles work when a connection is
attempted. Right now, the handles do not allow for new connections to be
made when they are not visible. Although this is intended, I need to
make the handles visible when the user is attempting to make a new
connection. I need to figure out how to do this based on the position of
the cursor. This will also help when I add functionality to the
connector nav item.