Added the ability to update node labels
Node labels can now be updated in the Attributes menu. When a node is
selected, the attributes menu will appear and when opened, will display
the current label of the node. The user can then edit the label and in
realtime update the current node.

There is a current bug that happens if the user holds shift/ctrl while
pressing delete/backspace when updating the label. If that happens, the
node will delete itself. This seems to be happening because of reactFlow's
useGlobalKeyHandler hook which does not check if the node is focused
before initiating the delete action.

In order to fix this however, I think I will have to either build the
ReactFlow component from scratch to remove theuseGlobalKeyHandler hook
or manually implement the delete functionality. I will have to look into
this more.