Removed unnecessary form for node updates
I was running into some weird issues with the form. I needed a useEffect
to update state of the form and there was some default behavior that
caused the page the refresh when return was pressed. Moving to a text
input will be simpler and easier to manage.

I was trying to make a simpler way of updating the node properties, but
realize that there isn't a way at compile time to know the type of node
that is being updated. I could simply have one node that has all the
various properties, but would need to change some things.

The idea would be that a single node would have all the properties and
simply have some of them hidden based on the type of node as a string.
From there, I could map through the properties that are visible and
editable so that I wouldn't have to hard code the accordion attributes.

I'll have to think about it a bit more. It's not top priority, but can
be a nice to have that simplifies the code a bit.