Removed other nodes in favor of one node
I thought a bit more about the previous commit and decided that it would
make sense to have one node that can be used for all the different types
of interactions. From the perspective of ReactFlow, the nodes were going
to all be identical anyways but simply have different types. By doing
things this way, I should be able to simplify the types significantly
and simply show or hide attributes and properties that are relevant to
the card shown.

I will determine what to render for each node based on a type that can
be found within the data prop of the ReactFlow node. That way, the
ReactFlow implementation is simpler and less likely to cause issues.

I will need to find out how to make the "card" node the default one, but
when I tried to override the default node type, both the card and the
default node were rendered.

The main thing now will be to configure the attributes such that when I
change the underlying data type, a differing set of properties will
appear to edit. The current setup hardcodes the properties, but I can
easily change this to be more dynamic. That way, I can have many different
types of cards, not have to update ReactFlow with custom nodes, and have
the attributes automatically handle the new card.