Adds the ability to change node status based on rule
The nodes originally had a status field that could be directly modified
via a checkbox. This has since been changed to now require a ruleId to
be specified in order to change the status. This is the first step in
setting up the nodes to be able to updated based on the user within the
testing section of the accordion.

Next steps are as follows:
- Add data to the testing section of the accordion and check rules based
on that data
- Add the ability to create new rules and set up a dropdown to select
the rule for each node
- Modify the colors and animation qualities of the edges based on the
status of the leading and trailing nodes

I've also added a new utility function to help with bool to status
conversions although I will probably need to rethink how I approach this
in the future. Statuses should probably be stored as enums and ruleIds
should probably not be nullable but instead default to having a rule of
"none".