Moved the accordion components into their own folders
Accordion components now exist in their own folders. I can probably name
them better, but for now, this works. I also added a new accordion
component for testing. This will be used to display the results of
the various flags and to test whether any individual flag is working
correctly.

The test component is not yet complete and only there for visuals.