Nodes now update base on flag rules
When user data changes within the testing section, nodes will update to
reflect the new data. This is done by using a custom hook that I created
alonside a useEffect that updates the nodes when the data changes.

It seems a bit clunky atm due to multiple places where there are
useEffects. Ideally, I would be able to use hooks to update the state of
the nodes. I may be able to get that working with the use of either the
builtin ReactFlow hooks or the use or a state management library like
Zustard which is recommended by the ReactFlow docs.

That being said, it's working! I can modify the data and the nodes will
update quickly. There is currently an issue around the date not working
which I think might have to do with the way they are being compared. I
will need to look into that.