Nodes with visible handles can connect to nodes
Before, unless both nodes were visible, they could not be connected. I
also had the issue where invisible handles could be the starting point
of a connection which I did not want.

This commit prevents invisible handles from starting an edge but allows
invisible handles to be the recipient of an edge.

I still need to figure out why they edges are so distant from the card
they are connected to.