Colors the edges based on parent node
This is a rather convoluted approach to getting the edges to be colored
and animated when the parent node is passing or failing.

The main issues at the moment are as follows:
- The edges update after the nodes which means that they are
occasionally out of sync. This can be seen when the rules are changed,
but nothing else is.
- Edges only care about their parent node. If that parent node is the
child of another node that is failing, the edge will still be colored
and animated as if it were passing.
- Edges and Nodes are stored as lists, which means that the edges need
to search through the list of nodes to find their parent. This is an
O(n^2) operation, which is extremely slow. This can heavily impact
performance when there are a large number of nodes and edges.
- The managing of state is a bit of a mess. Ideally there would be a
hook to manage the state of nodes and edges with some functions to add,
remove, edit, and update the nodes and edges. This would make the flow
diagram much more modular and easier to work with.

Although it would be drastic, I'm starting to lean towards looking into
other libraries to handle the flow diagram. The current library is fine,
but because of the issues stated above, it is not ideal.