Created a rules map for flag rules
This should make it easier to add, remove, modify, and search for rules
that are created. I still need to figure out how a user will be able to
create a rule, but this is a good start.

I'm also not a huge fan of forcing a rerender, but in doing it this way,
I don't have to create a new Map every time either the component is
rerendered, or when the rules are updated.