Created a cookie context with hook
There will be sections of the app that will need to know the current
cookie and potentially update it. Examples of this include the header
which contains the login/logout button, user details which includes the
user's name, and potentially some future content.

This makes it significantly easier to modify the cookie from anywhere