Adding some debug info on startup
I'd like to know that each of the services are running properly on
server startup. The issue is that although some of these services can
fail to initialize, the server will still continue to startup. I think
that I will need to move startup code to its own file and import it
within index.ts so that I can catch errors and exit the process if any
of the services fail to initialize.