Updating routing structure
This is a major update. I realized that with my previous commit, I
subtlely broke routing. It was actually an extremely confusing bug, and
I wasn't sure exactly why some of the routes were working and others
were not.

What was happening?
Routes like /api/user and /api/link/create were working, but routes like
/api/auth/login were not. I was confused because the code for the routes
was the same, but the routes were not resolving properly and returning
errors.

What was the problem?
The problem is two fold. The first thing to understand is how the router
works. Every single route has a path. This path is used to resolve the
route recursively. What that means is that I can have a route within a
route within another route and each of them only need to specify their
relative path.

Example: I can have an API router with a path "/api" that within it has
an Auth router with a path "/Auth" that within it has a login router
with a path "/login". Express is smart enough to realize that these
routes depend on each other and so when it creates the login route, it
knows to make the route "/api/Auth/login".

The second part is how I was creating the routes with the automated
route system. The idea is that I would create a similar folder structure
to NextJS where each route exists within a folder. Now, I just needed to
have a login.routes.ts file within an Auth folder that is within an api
folder for the route to be generated.

I was creating routes like so:
router.use("<pathToRoute>", <routerAtPath>)

This might at first glance seem fine. Take login as an example:
router.use("/api/auth/login", router);

But what's the issue here? Remember that each route has its own path?
This function is actually doing an additional recursive call appending
the path of the route to my constructed path.

Instead of the route being at "/api/auth/login", it was actually at
"/api/auth/login/login" because the login router had a path of "/login"!

That insidious error was causing my issues. Since the User and Link
routes were not changed since before my updates, they still existed in
the top level folder and had the names User and Link respectively.
They were being added as such:
router.use("/api/user", router);

And since they had routes on the "/" or "/update" directories, it just
worked because they would be added to the end properly.

So I definitely need testing as my next step. There is no doubt about
it. I at the very least created some simple conditionals to catch routes
with the same names and I have a hacky way to get routes (although I
should create a function that uses the same recursion as the actual
router) which should at the very least alert me of the types of routes
added, but man it would be great to be 100% sure that routes are done
right.

I've also migrated everything to the new route structure. Adding a new
route is super simple now.

On another note, I need to figure out how to either remove dist on every
server update or simply not emit JS code. Deleting files does not update
the dist folder which made this even more difficult to uncover.

On another nother note, I need to update how I run NodeJS. I can't get
paths like the Frontend working which results in a lot of relative
paths. It's frustrating to move files around and having to change all
dots over and over.