Added AJV validation to select routes
This commit adds a few things that I currently hadn't set up yet. This
involves having a start function (placeholder name) that initializes
everything that needs to be initialized. This will eventually house
stuff like prisma initialization, redis initialization, and more, but
for now, it has the validation compilation. This will cleanup some of
the middleware that isn't really middleware and allow me to stop the
server if initialization fails (not yet implemented).

I also created a way to automatically add all of the schema rules from
each of the routes into the ajv schema. This means that they don't have
to be added manually each time and they can exist only within the file
that they refer to. There are some issues such as the fact that it's
not robust since there isn't any typing or help within any of the route
modules to guide them on what fields are required. There is a simple
check that occurs on startup, but that seems rather weak. Ideally I
could type these routes in such a way that enforces what they need to
export in order to work.

I've also gone ahead and added uuid. This is used to track the routes
for ajv and to request the schema that is needed for each route. In an
ideal world, I could somehow automatically run validation for a route
based on whether or not they have some exported request interface, but
I need to think about it a bit more. For now, a uuid seems sufficient.