Seemingly fixed lint rule conflict
Because Prettier and Eslint have some opinions about how they want to
handle code formatting, I've decided to separate them. I don't entirely
understand how it works, but now lint errors should only be for code
quality and organization with styles existing in prettier.

I would like to somehow be able to have a configuration for this project
that I can use. I fear that it works for this system that I'm working on
but the moment I switch to another, it'll cause issues.

I also do like the idea of enforcing style choices similar to FB's
sandcastle, automating the styles, but it's way to much work for now.