Created initialization functions for services
Finally moved the initialization of middleware services to their own
startup file called initializate. This way, they can be easily organized
and handled.

There were some middleware functions that were running on application
start which confused me. I learned that whenever a file is imported in
JS, all of the constants are evaluated and called functions are run.
This meant that functions running for middleware had some stuff running
on application start and it was difficult to determine when they were
running in relation to other things. As a lesson, I will no longer have
functions that are called directly in a file as they can run at an
unexpected time.

Furthermore, I removed the router.ts file since it is no longer
necessary. The initializeRoutes function takes care of route creation
and schema generation for each route at the same time which is better
than before doing it separately (although performance differences are
negligible with such a small amount of routes).

One thing I would like to explore is how to only recompile files that I
have modified so that I don't have to reconnect to server and cache on
every save. Not sure how that would work with the express router, but I
think in a much later stage of my project it will be invaluable.

The next thing on my list will be to update my schemas to the new AJV
JSONSchema standard followed by adding some tests to ensure they are
configured properly.