Added ajv-errors and common schemas
With the use of ajv-errors, I can add custom errors for each of the
validation rules that I have set up. This allows me to split up the
rules into more manageable pieces and can give users more detailed
information when one of the rules fails. In contrast, only the regex
pattern was being displayed which isn't helpful.

I've also removed the old requestValidation file in favor of the new
validateRequest file. In a future commit I will also remove the extra
validation packages I was using too.

Next steps will be to modify the validation middleware so display errors
in a more human friendly format with a lot less information that I don't
want to reveal to those consuming my API.