Cleaned up errors and responses from server
When there is an error, the server will respond with an object
containing all of the errors for the given validation rules. Currently
there is only an error message that is passed to the user, but ideally
this would also include error codes. There will need to be consideration
for i18 standards in the future, but that's not something I want to even
think about for a while.