Modified nodemon in attempt to set up paths
I really struggled attempting to get paths working. I tried the
following:
 - Webpack
 - Babel
 - Rollup
 - tsconfig-paths
 - Tried screaming and crying
 - module-alias
 - Switching between module and commonJS
 - Tried logging through tsconfig-paths package to see if it was working

Nothing worked. I don't know why. At this point, I have given up on it.
This is how it will be for the forseable future until I find someone to
help me because FUCK. I didn't think this would be so difficult. I am
just guessing and checking and nothing works. I've followed guides,
literally copying and pasting what others have done and no luck.

My best guess is that this isn't working due to the nature of the
monorepo, but I can't for the life of me figure out why. It has to be
something simple, but I'm not going to waste more time on this.

I switched to a different nodemon script because I figured it would be
better to use rimraf to remove the old dist and create it again and
remove the expirimental es module resolution field. This meant I had to
go through every single file and add the .js extension which sucked. But
because I was changing settings and reverting, that happened 5 times. So
that sucked x5. At least now it matches ECMAScript standards for
relative imports, but if I get paths working, I just throw that out the
window again huh.

Anyways, fuck this. I ain't doing paths anytime soon.