Adding Jest to project with a test case
Adds Jest to the project and prevents tests from becoming part of dist.
Apparently you can make the jest.config.js file into a TS file, but I
wasn't able to do that even with adding "jest.config.ts" to the includes
field within tsconfig.json.

Luckily this isn't an issue because I can use a comment to add in the
types and jest will report back with type errors regardless so I'm not
going to focus on that anymore.

This is a great step though! It can actually find the tests that are
nested inside of my hierarchy which is great! I don't even have to add
any custom regex rules besides the default ones. I may take some time
to comb through all of the jest features to see if there is anything
that might make the tests better, but that is for a later commit.