Added supertest and a jest test case
Had to switch back to the previous nodemon command + rimraf. I was
running into issues with jest not being able to resolve files with the
js extension and figure that I should deal with that later. For now, I
was able to create a custom route to test with jest and supertest.

Supertest is creating a custom request that can take in an express app
and handle the route passed to it. This allows for me to do unit testing
on each route that I have. I haven't used an actual route yet, because I
need to mock some of the middleware that exists, namely session and
prisma, but if I move the authentication to a middleware (isLoggedIn),
then I may just have to mock only prisma.

Regardless, this is a good step! I shuold have a test for at least one
route shortly.