Modified tsconfig and build to support types
With the previous version of my tsconfig, I excluded jest because doing
so prevented the tests from appearing in the dist folder. This is useful
because we don't need tests in the production environment. As a
consequence of this however, custom types that I created for handling
sessions was also being ignored and causing jest issues with new tests.
I have since created a special tsconfig specifically for builds which
excludes tests on build. So now, jest TS files gain the linting, and
type resolution, while the build version ignores jest tests entirely.