Added a test case for /api/user/
Added two middleware mocks that allow for the /api/user/ route to be
tested. I also added the jest-mock-extended library which makes mocking
stuff like PrismaClient significantly easier as it does not require that
I manually type everything as jest.fn().

Although this works, I think I should change up my middleware a bit. The
main issue is that I am creating a set of custom middleware options that
does not reflect the actual application. They are prone to setting in
the wrong order and potentially giving bad results. The best approach
will be to set up my middleware such that it had dependency inversion.
I already check if the values for stuff that prisma and router are valid
before allowing applyMiddleware to run, but I should instead pass those
values in to applyMiddleware so that they don't just "hope" they are
valid. This also has the benefit of providing me with a way to get the
actual middleware implementation and pass in mocked values like I have
in this current test. It will make them more reliable and should make my
code more robust.