Organizing middleware and startup
I was trying to add tests and realized that I couldn't mock any of the
clients that I was interacting with (Prisma, Redis) so I moved the
initialization of the startup around so that I could pass mocked data to
my middleware. Now I can have mocked fields for my middleware that I can
setup for testing.

The issue however is that this feels like an integration test rather
than a unit test. I'm testing the entire process from app startup to
response rather than crafting a request from scratch. I've also not been
able to totally mock the redis store for my session middleware to be
properly mocked either. This has resulted in not properly storing
session data in the request.

As such, I've decided that although tests are failing for this commit,
I am going to push and create actual unit tests for this route. In the
future, I will look to run my routes through some proper integration
tests.