Created test cases for login
Learned about jest.spyOn and how it is necessary for replacing function
implementations across other files. I've also gone ahead and removed the
autotyping from "node" as it was getting annoying typing things. Node
and Express share the same type names and it was auto resolving to node.

I moved the ParamsDictionary type into the types.d.ts file as it was not
autopopulating from the express dependency library. I think this is
better for now until they change the type in which case I will come back
to it.

I added in the jest.config.js file a new property that will reset mocks
between tests. Although this doesn't work for mocked objects, this is
great for mocked functions so that jest resolvers don't affect other
tests accidentally.