Added testing for register
I believe that when there is an error, I should probably only send back
the error rather than include a null field or something similar. This is
the approach that I took with this route, but I'm still on the fence
about it.

I also added in that when registration is complete, the user is
automatically logged in for convenience.