Modified initialization functions
I was trying to make initliazeRoutes testable, but was running into tons
of issues trying to mock fs, path, and other node modules. Instead, I
decided to skip that for now. Each of the initialize functions will have
an errors array that is read. Failures are still not handled so I'll do
that on another diff as well.