Adding connection to PlanetScale
I'm trying to figure out whether I should stick with prisma or switch to
planetscale. I know they are advertised together, but I think that it
will be easier to build business logic with an SQL server rather than
using Prisma directly.

The lack of type safety sucks, but I think it will make more sense in
the long run.