Added SignIn component with a form
I've built this component in another commit but was not able to push it.
I've rebuilt the component and will next connect it to my backend.

I've updated my cors policy to accept connections from my frontend. I
will need to look into how nginx works so that I can use a proxy to host
my backend and frontend on the same domain.