Added Axios and some name changes
I added axios to the frontend to handle requests to the backend. I was
debating between using axios and fetch and have decided that axios is
easier to implement and has some additional features that may be useful
in the future.

I also changed a directory from util -> utils to be more consistent with
the rest of the project.