Added form loading and updated error message
The signin form will now properly start loading when the request is in
progress. I've updated the error message on the login route to be less
descriptive to the user. I'm not sure if it helps to tell the user that
there is not an email in the database on login.

I've also added a redirect from the signin page to the dashboard when
sign in is successful.

Notes:
I think that in the future, I would like to type all of my axios
requests. Although I can do that now by referencing the types within the
backend project, I think it will be better to create a codegen that uses
swagger to generate those types. My reasoning is that it another dev
tries to consume my api, they will be doing the same thing. This should
ensure that my docs are up to date. I will do some more research about
it in the future.